requirements.txt
fastapi
uvicorn
requests
slowapi
duckduckgo-search
google-generativeai
python-dotenv